Blind decryption shares such that they're only usable by the decryptor
This is posited as a t-of-n work which means there are non-participating
signers. If the shares are public, one more share is needed to complete the
decryption. While this is expected to be the first signer, it could also have
been a non-participating signer. This non-participating signer, being the sole
malicious participant, could have also performed decryption (recovering `d`)
and from there recovered the first signer's private key from their key share.

This blinding prevents such behavior.