Calculcate all of w in-ciphertext
- Removes a proof column for Y_i, as prior needed for verifying shares

We prior used an entire additional proof just for this.

- Removes subtractive shares

- Removes participant asymmetry

- Removes designed-decryptor decryption shares (reducing the size of their
  proofs)

- Enables delaying verification of the decryption share proofs until an invalid
  signature is generated

Prior, only one party received the signature, so a lack of signature
publication to the group was indistinguishable due to abort or the recipient
going offline. Now, everyone gets the signature, and anyone realizing they
produced an invalid signature can verify the decryption proofs.